requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.9
langchain-google-genai>=1.0.6
yt-dlp>=2024.5.27
opencv-python>=4.10.0.82
openai-whisper>=20231117
pytesseract>=0.3.10
pydantic-settings>=2.3.4
python-dotenv>=1.0.1
google-generativeai>=0.5.2,<0.6.0
openai>=1.35.3
duckduckgo-search>=6.1.7
asyncio>=3.4.3
numpy>=1.26.4
torch>=2.3.1
pillow>=10.3.0
ffmpeg-python>=0.2.0
